Consolidate js into one file, and defer loading
