+++ web/templates/footer.php
@@ -55,5 +55,37 @@
 		<span class="u-hidden"><?=_('Top');?></span>
 	</button>
 
+<?php
+	if (!empty($_SESSION['error_msg'])):
+	?>
+	<div>
+		<script>
+			$(function() {
+				$('#dialog:ui-dialog').dialog('destroy');
+				$('#dialog-message').dialog({
+					modal: true,
+					resizable: false,
+					buttons: {
+						Ok: function() {
+							$(this).dialog('close');
+						}
+					},
+					create: function() {
+						var buttonGroup = $(this).closest(".ui-dialog").find('.ui-dialog-buttonset');
+						buttonGroup.find('button:first').addClass('button submit')
+						buttonGroup.find('button:last').addClass('button button-secondary cancel');
+					}
+				});
+			});
+		</script>
+		<div id="dialog-message" title="">
+			<p><?=htmlentities($_SESSION['error_msg'])?></p>
+		</div>
+	</div>
+<?php
+	unset($_SESSION['error_msg']);
+	endif;
+?>
+
 </body>
 </html>
