+++ web/templates/header.php
@@ -5,20 +5,8 @@
 <?php
 	require $_SERVER['HESTIA'] . '/web/templates/includes/title.php';
 	require $_SERVER['HESTIA'] . '/web/templates/includes/css.php';
-	require $_SERVER['HESTIA'] . '/web/templates/includes/top_js.php';
+	require $_SERVER['HESTIA'] . '/web/templates/includes/js.php';
 ?>
-	<script>
-<?php
-		//
-		//	GLOBAL SETTINGS
-		//
-?>
-		var GLOBAL = {};
-		GLOBAL.FTP_USER_PREFIX = '';
-		GLOBAL.DB_USER_PREFIX = '';
-		GLOBAL.DB_DBNAME_PREFIX = '';
-		GLOBAL.AJAX_URL = '';
-	</script>
 </head>
 
 <body class="body-<?=strtolower($TAB)?> lang-<?=$_SESSION['language']?>">
