--- web/templates/includes/top_js.php
@@ -1,1 +1,0 @@
-<script src="/js/vendor/jquery-3.6.1.min.js"></script>
